updated styling for intro, tech and header
