+++ src/components/Intro/Intro.js
@@ -1,17 +1,17 @@
 /* eslint-disable max-len */
 import React from 'react';
-import styled from 'styled-components';
+import styled from 'styled-components/macro';
 
-import { OuterWrapper } from '../Styles/Globalstyles';
+import { OuterWrapper, InnerWrapper } from '../Styles/Globalstyles';
 
 const Intro = () => {
   return (
     <StyledIntro>
-      <IntroText>
+      <InnerWrapper>
         <Line /><Dot />Hi! I am Julia,  a creative-minded team player with a love for new technologies and design.
         I strive to constantly grow while solving problems that matter and creating engaging solutions for products that people love!
         I stay motivated by collaborating with like-minded and with a curiosity about learning new things.
-      </IntroText>
+      </InnerWrapper>
     </StyledIntro>
   )
 }
@@ -19,26 +19,8 @@ const Intro = () => {
 export default Intro
 
 const StyledIntro = styled(OuterWrapper)`
-  min-height: 25vh;
   background-color: var(--color-grey);
-
-  @media screen and (min-width: 667px) {
-    min-height: 20vh;
-  }
   `
-
-const IntroText = styled.div`
-  max-width: 80vw;
-  line-height: 1.3rem;
-
-  @media (min-width: 667px) and (max-width: 1024px) {
-    width: 60vw;
-  }
-
-  @media screen and (min-width: 1024px) {
-    width: 40vw;
-  }
-`
 
 const Line = styled.span`
   border: 1px solid var(--color-darkGreen);
