+++ src/components/MoreThoughts.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import styled from 'styled-components'
+import styled from 'styled-components/macro'
 
 import { SectionSubTitle, OuterWrapper } from './Styles/Globalstyles'
